Moved update logic to AdminSubscriber.php
